requirements.txt
# Replicate Lipsync API for ComfyUI - Requirements

# Core dependencies
replicate>=0.25.0
requests>=2.25.0
python-dotenv>=1.0.0
numpy>=1.19.0
opencv-python>=4.5.0

# Audio processing (required for audio input support)
soundfile>=0.10.0
scipy>=1.7.0

# Optional: For advanced features
# torch>=1.9.0  # Usually provided by ComfyUI
# Pillow>=8.0.0  # Usually provided by ComfyUI
